IDP-1921 refactor qrcode preauth
